+++ docs/guides/changelog.md
@@ -10,6 +10,7 @@ description: |-
 ## Unreleased
 
 - (Enhancement) Performance improvements
+- (Enhancement) `fmc_device`: Add `access_control_policy_domain`, `nat_policy_domain` and `health_policy_domain` to support policies that exist in a different domain than the device
 - (Enhancement) `fmc_device_physical_interface`: Increase the maximum MTU from `9000` to `9198`
 - (Enhancement) `fmc_device_etherchannel_interface`: Increase the maximum MTU from `9000` to `9198`
 - (Enhancement) `fmc_device_subinterface`: Increase the maximum MTU from `9000` to `9198`